requirements.txt
grpcio
grpcio-tools
requests
pymongo
protobuf
python-dotenv
Flask
pytest